requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
sqlite-utils==3.36
apscheduler==3.10.4
twilio==9.0.4
reportlab==4.2.0
python-dotenv==1.0.1
websockets==12.0
Pillow==10.3.0
python-multipart==0.0.9
httpx==0.27.0
